backend/requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlmodel==0.0.22
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.2.0
python-dotenv==1.0.1

# Validation and Data
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.3.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Code Quality
ruff==0.8.5
mypy==1.14.1

# CORS
fastapi-cors==0.0.6

# AI and MCP
google-generativeai==0.8.3
mcp[cli]==1.25.0
slowapi==0.1.9
